refactor(routes): moves auth routes into actions
